requirements.txt
pytest~=8.3.3
fastapi~=0.115.12
email-validator
python-dotenv
requests~=2.32.3
pydantic~=2.11.5
jsonschema~=4.23.0
schemas~=0.7.1
uvicorn~=0.34.2
fastapi-pagination
psycopg2-binary
sqlalchemy
sqlmodel
fastapi[standard]







